requirements.txt
python-telegram-bot>=20.0
APScheduler>=3.10.1 